*made the permission prices as doubles;

+++ src/de/xzise/wrappers/economy/AccountWrapper.java
@@ -6,7 +6,7 @@ package de.xzise.wrappers.economy;
  */
 public interface AccountWrapper {
 
-    boolean hasEnough(int price);
-    void add(int price);
+    boolean hasEnough(double price);
+    void add(double price);
     
 }
