+++ src/de/xzise/wrappers/economy/EconomyWrapper.java
@@ -7,7 +7,7 @@ import de.xzise.wrappers.Wrapper;
 public interface EconomyWrapper extends Wrapper {
 
     AccountWrapper getAccount(String name);
-    String format(int price);
+    String format(double price);
     
     Plugin getPlugin();
